try to track l10n status

+++ lib/libtag.rb
@@ -40,24 +40,26 @@ end
 # * Svn cp from fetched translations (libl10n)
 # TODO: optionalify depend on libl10n
 def tagTranslations()
-    @name = NAME.split("-").join
+    unless @translations == nil
+        @name = NAME.split("-").join
 
-    Dir.chdir( BASEPATH + "/" + @folder )
-    `svn co -N #{@tag1} tagging`
+        Dir.chdir( BASEPATH + "/" + @folder )
+        `svn co -N #{@tag1} tagging`
 
-    tag = "#{@protocol}://#{@user}.kde.org/home/kde/tags/#{NAME}/#{@version}/po"
+        tag = "#{@protocol}://#{@user}.kde.org/home/kde/tags/#{NAME}/#{@version}/po"
 
-    `svn mkdir -m "Create tag #{NAME} #{@version} po directory" #{tag}`
-    `svn up tagging/po`
-    for translation in @translations do
-        `svn mkdir tagging/po/#{translation}`
-        for f in Dir.glob("po/#{translation}/#{@name.chop}*.po")
-            `svn cp #{f} tagging/po/#{translation}/`
+        `svn mkdir -m "Create tag #{NAME} #{@version} po directory" #{tag}`
+        `svn up tagging/po`
+        for translation in @translations do
+            `svn mkdir tagging/po/#{translation}`
+            for f in Dir.glob("po/#{translation}/#{@name.chop}*.po")
+                `svn cp #{f} tagging/po/#{translation}/`
+            end
         end
-    end
-    `svn ci -m "Tag #{NAME} #{@version} - translations." tagging/po`
+        `svn ci -m "Tag #{NAME} #{@version} - translations." tagging/po`
 
-    FileUtils.rm_rf("tagging")
+        FileUtils.rm_rf("tagging")
+    end
 end
 
 # Tags documentation - 3rd step
@@ -66,19 +68,21 @@ end
 # * Svn mkdir doc
 # * Svn cp DOC for all DOCS (provided by libl10n. So, if no translation fetching did happen, it's going o break here)
 def tagDocumentations()
-    Dir.chdir( BASEPATH + "/" + @folder )
-    `svn co -N #{@tag1} tagging`
+    unless @docs == nil
+        Dir.chdir( BASEPATH + "/" + @folder )
+        `svn co -N #{@tag1} tagging`
 
-    tag = "#{@protocol}://#{@user}.kde.org/home/kde/tags/#{NAME}/#{@version}/po"
+        tag = "#{@protocol}://#{@user}.kde.org/home/kde/tags/#{NAME}/#{@version}/po"
 
-    `svn mkdir -m "Create tag #{NAME} #{@version} doc directory" #{tag}`
-    `svn up tagging/doc`
-    for doc in @docs do
-        `svn cp doc/#{doc} tagging/doc/`
+        `svn mkdir -m "Create tag #{NAME} #{@version} doc directory" #{tag}`
+        `svn up tagging/doc`
+        for doc in @docs do
+            `svn cp doc/#{doc} tagging/doc/`
+        end
+        `svn ci -m "Tag #{NAME} #{@version} - documentations." tagging/doc`
+
+        FileUtils.rm_rf( "tagging" )
     end
-    `svn ci -m "Tag #{NAME} #{@version} - documentations." tagging/doc`
-
-    FileUtils.rm_rf( "tagging" )
 end
 
 # Tagging wrapper
